1.ffl基础库支持android 2.修改播放器lib

+++ ffl/include/utils/FFL_Clock.hpp
@@ -81,7 +81,7 @@ namespace FFL {
 		//
 		//  y=ax + b + offset
 		//
-        int64_t mA;
+        int32_t mA;
         int64_t mB;
         //
         // 系统时间偏移值
